Revert "Server client communicction"
