+++ Projects/MakeMeLaugh_Client/Assets/Scripts/JokeTempalte.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0375d7b884656ca448c5e44d5f220bf0
+guid: ebde38b58e69ac94486285f40cf78af5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
