--- Projects/MakeMeLaugh_Server/Assets/Scripts/IPManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b1b74505ba6213740a5e2eeeb1432152
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
