--- SharedPackages/Assets/Scripts/JokeState.cs
@@ -1,7 +1,0 @@
-﻿public enum JokeState
-{
-  Setup,
-  Punchline,
-  Done,
-}
-
