--- SharedPackages/Assets/Scripts/PlayerPunchlineRequest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e509c87049d482e4a8a2378c745a3f49
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
