--- SharedPackages/Assets/Scripts/PlayerPunchlineResponse.cs
@@ -1,11 +1,0 @@
-﻿public struct PlayerPunchlineResponse
-{
-  public PlayerPunchlineResponse(string punchlineSegment, string jokeId)
-  {
-    PunchlineSegment = punchlineSegment;
-    JokeId = jokeId;
-  }
-  public string PunchlineSegment;
-  public string JokeId;
-}
-
