--- SharedPackages/Assets/Scripts/PlayerSetupRequest.cs
@@ -1,11 +1,0 @@
-﻿public struct PlayerSetupRequest
-{
-  public PlayerSetupRequest(string setupTemplate, string jokeId)
-  {
-    SetupTemplate = setupTemplate;
-    JokeId = jokeId;
-  }
-  public string SetupTemplate;
-  public string JokeId;
-}
-
