--- SharedPackages/Assets/Scripts/PlayerSetupRequest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4b6d2bbc00df5064aa3ecf6d4503f7c5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
