--- SharedPackages/Assets/Scripts/PlayerSetupResponse.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: df68f124eec16f74c9e89bd0dbc48596
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
